Instalar e configurar eslint e babel-root-import
Foi instalado e configurado o eslint baseado na style guide do Airbnb,
sobrescrevendo algumas regras. Também foi instalado o
babel-plugin-root-import para configurar o src/ como diretório base
para realizar as importações. Foi adicionado um .editorconfig para
definir as mesmas regras de formatação em todos os editores de texto.
Também foi criado um jsconfig.json para habilitar o intellisense do
babel-plugin-root-import ao VS Code.
