+++ babel.config.js
@@ -1,3 +1,12 @@
 module.exports = {
-  presets: ['module:metro-react-native-babel-preset'],
-};
+  presets: [ 'module:metro-react-native-babel-preset' ],
+  plugins: [
+    [
+      'babel-plugin-root-import',
+      {
+        rootPathPrefix: '~',
+        rootPathSuffix: 'src',
+      },
+    ],
+  ],
+}
