+++ src/components/test.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Text } from 'react-native'
+
+const Test = () => (
+  <Text>Test component</Text>
+)
+
+export default Test
