+++ src/pages/page.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Text } from 'react-native'
+import Test from '~/components/test'
+
+const Page = () => (
+  <>
+    <Text>Page</Text>
+    <Test />
+  </>
+)
+
+export default Page
